refactor(lode): ♻️ decouple lode from cli/reader dependency
Move ParseMetricsRecord and its helpers into cli/reader where the
MetricsSnapshot type already lives. This eliminates the layering
violation where the storage package (lode/) imported a CLI
presentation package (cli/reader/), unblocking the future
core/cli/temporal module split.

- New file: cli/reader/parse.go (ParseMetricsRecord + conversion helpers)
- New file: cli/reader/parse_test.go (4 unit tests moved from lode/)
- Rewrite lode/ round-trip tests to assert on raw map[string]any
- Update stats.go call site: lode.ParseMetricsRecord → reader.ParseMetricsRecord
